gluon-core, gluon-wan-dnsmasq: also consider 'cellular' an uplink interface
QMI interfaces will not always create a separate 'cellular_6' interface
for IPv6. Also accept configuration found directly on the 'cellular'
interface to fix IPv6 DNS in this case.

The firewall rules are adjusted for consistency, although here the issue
appears to be cosmetic, as the 'cellular_4' interface results in wwan0
rules to be added to ip6tables as well.